refactor: move error middleware to middleware dir
